+++ vite.config.ts
@@ -4,5 +4,5 @@ import react from '@vitejs/plugin-react'
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
-  base: '/ScriptureFlux/', // GitHub Pages 배포 서브디렉토리 지원
+  base: '/scripture-flux-web/', // GitHub Pages 배포 서브디렉토리 지원
 })